Created boilerplate for sprint3.md doc
Copied and pasted from Sprint2.md to start. I cleared the completed work section, so we'll need to fill that out. I added in frontend unit tests to-date and removed the cypress section. I did not change/remove any of the contents in the backend section, so that will need to be updated.